refactor: move pre-flight checks and run queue from infra to zero layer
- Extract checkRunConcurrencyLimit, checkOrgCredits, checkModelProviderConfigured
  to zero-preflight.ts (zero layer owns business logic)
- Move queue management (enqueueZeroRun, drainOrgQueue, drainStaleQueues,
  cleanupExpiredQueueEntries) to zero-queue-service.ts
- Delete run-queue-service.ts (infra layer no longer manages queues)
- Add existingRunId to CreateRunParams: createRunRecord UPDATEs when set,
  INSERTs otherwise, unifying direct and dequeued run paths
- Replace queueDispatcher param with drain callback on buildAndDispatchRun
  and dispatchCancelSideEffects
- startRun() calls pre-flight checks from zero-preflight.ts before
  delegating to createRun, maintaining behavior for non-zero API routes
- createRun() is now a pure execution function (no pre-flight checks)
- Move credit-check and queue tests from run/ to zero/ test directories

Closes #7203
