+++ turbo/apps/web/app/api/agent/runs/[id]/cancel/route.ts
@@ -14,7 +14,7 @@ import {
   cancelRun,
   dispatchCancelSideEffects,
 } from "../../../../../../src/lib/run/run-service";
-import { dispatchQueuedZeroRun } from "../../../../../../src/lib/zero/zero-queue-service";
+import { drainOrgQueue } from "../../../../../../src/lib/zero/zero-queue-service";
 import { isNotFound, isBadRequest } from "../../../../../../src/lib/errors";
 import { logger } from "../../../../../../src/lib/logger";
 import { after } from "next/server";
@@ -61,7 +61,9 @@ const router = tsr.router(runsCancelContract, {
     try {
       const result = await cancelRun(runId, userId, orgId);
 
-      after(() => dispatchCancelSideEffects(result, dispatchQueuedZeroRun));
+      after(() =>
+        dispatchCancelSideEffects(result, () => drainOrgQueue(result.orgId)),
+      );
 
       log.debug(
         `Run ${runId} cancelled by user ${userId}, sandbox: ${result.sandboxId ?? "none"}`,
