+++ turbo/apps/web/app/api/cron/cleanup-sandboxes/route.ts
@@ -17,8 +17,7 @@ import {
   cleanupExpiredQueueEntries,
   drainStaleQueues,
   drainOrgQueue,
-} from "../../../../src/lib/run/run-queue-service";
-import { dispatchQueuedZeroRun } from "../../../../src/lib/zero/zero-queue-service";
+} from "../../../../src/lib/zero/zero-queue-service";
 import { processOrgCredits } from "../../../../src/lib/credit/credit-service";
 import { logger } from "../../../../src/lib/logger";
 import { env } from "../../../../src/env";
@@ -189,7 +188,7 @@ const router = tsr.router(cronCleanupSandboxesContract, {
     // as it serves as the fallback for missed webhook-triggered drains.
     const [expiredQueueCount, drainedCount] = await Promise.all([
       cleanupExpiredQueueEntries(),
-      drainStaleQueues(dispatchQueuedZeroRun),
+      drainStaleQueues(),
     ]);
 
     if (expiredQueueCount > 0 || drainedCount > 0) {
@@ -234,7 +233,7 @@ const router = tsr.router(cronCleanupSandboxesContract, {
             run.id,
             "timeout",
             timeoutReason,
-            () => drainOrgQueue(run.orgId, dispatchQueuedZeroRun),
+            () => drainOrgQueue(run.orgId),
           );
 
           await processOrgCredits(run.orgId);
