+++ turbo/apps/web/app/api/webhooks/agent/complete/route.ts
@@ -20,8 +20,7 @@ import type {
 } from "../../../../../src/lib/checkpoint";
 import type { RunResult } from "../../../../../src/lib/run/types";
 import { logger } from "../../../../../src/lib/logger";
-import { drainOrgQueue } from "../../../../../src/lib/run/run-queue-service";
-import { dispatchQueuedZeroRun } from "../../../../../src/lib/zero/zero-queue-service";
+import { drainOrgQueue } from "../../../../../src/lib/zero/zero-queue-service";
 import { processOrgCredits } from "../../../../../src/lib/credit/credit-service";
 import { appendChatMessages } from "../../../../../src/lib/agent-session/agent-session-service";
 import { chatThreadRuns } from "../../../../../src/db/schema/chat-thread";
@@ -185,7 +184,7 @@ function scheduleTerminalSideEffects(
 ): void {
   after(async () => {
     await dispatchTerminalSideEffects(runId, status, errorMsg, () =>
-      drainOrgQueue(orgId, dispatchQueuedZeroRun),
+      drainOrgQueue(orgId),
     );
     await processOrgCredits(orgId);
   });
