+++ turbo/apps/web/app/api/zero/runs/[id]/cancel/route.ts
@@ -14,7 +14,7 @@ import {
   cancelRun,
   dispatchCancelSideEffects,
 } from "../../../../../../src/lib/run/run-service";
-import { dispatchQueuedZeroRun } from "../../../../../../src/lib/zero/zero-queue-service";
+import { drainOrgQueue } from "../../../../../../src/lib/zero/zero-queue-service";
 import { isNotFound, isBadRequest } from "../../../../../../src/lib/errors";
 import { after } from "next/server";
 
@@ -34,7 +34,9 @@ const router = tsr.router(zeroRunsCancelContract, {
     try {
       const result = await cancelRun(params.id, userId, org.orgId);
 
-      after(() => dispatchCancelSideEffects(result, dispatchQueuedZeroRun));
+      after(() =>
+        dispatchCancelSideEffects(result, () => drainOrgQueue(result.orgId)),
+      );
 
       return {
         status: 200 as const,
