+++ turbo/apps/web/src/lib/run/index.ts
@@ -6,6 +6,7 @@
 export {
   validateAgentSession,
   startRun,
+  createRun,
   createRunRecord,
   buildAndDispatchRun,
   resolveStartRunCompose,
