--- turbo/apps/web/src/lib/run/run-queue-service.ts
@@ -1,418 +1,0 @@
-import { eq, lt, and, or, count, gt, sql, inArray } from "drizzle-orm";
-import { agentRuns } from "../../db/schema/agent-run";
-import { transitionRunStatus } from "./run-status";
-import { agentRunQueue } from "../../db/schema/agent-run-queue";
-import { orgMetadata } from "../../db/schema/org-metadata";
-import { modelProviders } from "../../db/schema/model-provider";
-import { env } from "../../env";
-import { logger } from "../logger";
-import { isConcurrentRunLimit, insufficientCredits } from "../errors";
-import { ORG_SENTINEL_USER_ID } from "../org/org-sentinel";
-import {
-  encryptSecretsMap,
-  decryptSecretsMap,
-} from "../crypto/secrets-encryption";
-import { PENDING_RUN_TTL_MS, checkRunConcurrencyLimit } from "./run-service";
-import type { CreateRunParams, CreateRunResult } from "./run-service";
-import type { OrgTier } from "@vm0/core";
-
-const log = logger("service:run-queue");
-
-// Queue entry TTL: 2 hours (same as runner_job_queue)
-const QUEUE_TTL_MS = 2 * 60 * 60 * 1000;
-
-/**
- * Dispatcher function type for queued runs.
- * Receives a run that has already been dequeued and transitioned to "pending".
- * Injected by callers to avoid circular dependency with run-service.
- */
-type QueuedRunDispatcher = (
-  runId: string,
-  createdAt: Date,
-  params: CreateRunParams,
-) => Promise<void>;
-
-/**
- * Enqueue a run that hit the concurrency limit.
- *
- * Creates a visible agent_runs record (status="queued") and stores
- * the full CreateRunParams in agent_run_queue with AES-256-GCM encryption.
- * The queue entry is deleted on dequeue — secrets never persist long-term.
- */
-export async function enqueueRun(
-  params: CreateRunParams,
-): Promise<CreateRunResult> {
-  const { userId, agentComposeVersionId, prompt } = params;
-
-  // Org context is required from caller
-  const orgId = params.orgId;
-
-  // Encrypt the full CreateRunParams for later replay
-  const paramsJson = JSON.stringify(params);
-  const encryptedParams = encryptSecretsMap(
-    { __params: paramsJson },
-    env().SECRETS_ENCRYPTION_KEY,
-  );
-
-  // Insert agent_runs + queue entry atomically to prevent orphaned records
-  const expiresAt = new Date(Date.now() + QUEUE_TTL_MS);
-
-  const run = await globalThis.services.db.transaction(async (tx) => {
-    const [inserted] = await tx
-      .insert(agentRuns)
-      .values({
-        userId,
-        orgId,
-        agentComposeVersionId,
-        status: "queued",
-        prompt,
-        appendSystemPrompt: params.appendSystemPrompt ?? null,
-        vars: params.vars ?? null,
-        secretNames: params.secrets ? Object.keys(params.secrets) : null,
-        resumedFromCheckpointId: params.resumedFromCheckpointId ?? null,
-        continuedFromSessionId: params.sessionId ?? null,
-        modelProvider: params.modelProvider ?? null,
-        lastHeartbeatAt: new Date(),
-      })
-      .returning();
-
-    if (!inserted) {
-      throw new Error("Failed to create queued run record");
-    }
-
-    await tx.insert(agentRunQueue).values({
-      runId: inserted.id,
-      userId,
-      orgId,
-      encryptedParams,
-      createdAt: inserted.createdAt,
-      expiresAt,
-    });
-
-    return inserted;
-  });
-
-  log.debug(`Enqueued run ${run.id} for user ${userId}`);
-
-  return {
-    runId: run.id,
-    status: "queued",
-    createdAt: run.createdAt,
-  };
-}
-
-/**
- * Drain the run queue for an org.
- *
- * Atomically dequeues the oldest entry, checks concurrency, deletes the
- * queue record, and transitions the run to "pending" — all within a single
- * advisory-locked transaction. This eliminates the orphaned run risk that
- * existed when dequeue and execute were separate transactions.
- *
- * Uses an iterative approach: on dispatch failure, marks the run as failed
- * and tries the next entry. Stops when a run is successfully dispatched,
- * the queue is empty, or the concurrency limit is reached.
- *
- * Called from:
- * - Completion webhook (event-driven, primary path)
- * - Cancel handler (when cancelling frees a slot)
- * - Cleanup cron (fallback for missed dequeues)
- *
- * @param orgId - Org whose queue to drain
- * @param dispatch - Dispatcher function (injected to avoid circular dependency)
- */
-export async function drainOrgQueue(
-  orgId: string,
-  dispatch: QueuedRunDispatcher,
-): Promise<void> {
-  const db = globalThis.services.db;
-  const encryptionKey = env().SECRETS_ENCRYPTION_KEY;
-
-  for (;;) {
-    // Single transaction: advisory lock → concurrency check → dequeue → status update
-    const dequeued = await dequeueNextAtomic(db, orgId);
-    if (!dequeued) return; // Queue empty, entry skipped, or concurrency full
-
-    // Decrypt CreateRunParams (outside transaction — no lock held)
-    const decryptedMap = decryptSecretsMap(
-      dequeued.encryptedParams,
-      encryptionKey,
-    );
-    if (!decryptedMap?.__params) {
-      log.error(`Failed to decrypt params for queued run ${dequeued.runId}`);
-      await markQueuedRunFailed(
-        dequeued.runId,
-        "Failed to decrypt queued run params",
-      );
-      continue; // Try next entry
-    }
-
-    const params: CreateRunParams = JSON.parse(decryptedMap.__params);
-
-    // Dispatch the run (compose loading, authorization, execution)
-    try {
-      await dispatch(dequeued.runId, dequeued.createdAt, params);
-      log.debug(`Queued run ${dequeued.runId} dispatched successfully`);
-      return; // Successfully dispatched — done
-    } catch (error) {
-      const errorMessage =
-        error instanceof Error ? error.message : "Unknown error";
-      log.error(
-        `Failed to dispatch queued run ${dequeued.runId}: ${errorMessage}`,
-      );
-      await markQueuedRunFailed(dequeued.runId, errorMessage);
-      continue; // Try next entry
-    }
-  }
-}
-
-interface DequeuedEntry {
-  runId: string;
-  createdAt: Date;
-  encryptedParams: string | null;
-}
-
-/**
- * Atomically dequeue the oldest queue entry for an org.
- *
- * Within a single advisory-locked transaction:
- * 1. Check concurrency limit (don't dequeue if no slot)
- * 2. Select oldest queue entry
- * 3. Delete queue entry
- * 4. Transition run from "queued" to "pending"
- *
- * If a run was already processed (e.g. cancelled), its queue entry is
- * removed and the next entry is tried within the same transaction.
- *
- * Returns undefined if queue is empty or concurrency limit reached.
- */
-async function dequeueNextAtomic(
-  db: typeof globalThis.services.db,
-  orgId: string,
-): Promise<DequeuedEntry | undefined> {
-  try {
-    return await db.transaction(async (tx) => {
-      // Serialize all queue operations for this org
-      await tx.execute(sql`SELECT pg_advisory_xact_lock(hashtext(${orgId}))`);
-
-      // Look up org tier and credits from org table (source of truth).
-      // Falls back to "free" (most conservative limit) if row is missing.
-      const [orgRow] = await tx
-        .select({ tier: orgMetadata.tier, credits: orgMetadata.credits })
-        .from(orgMetadata)
-        .where(eq(orgMetadata.orgId, orgId))
-        .limit(1);
-      const orgTier = parseOrgTier(orgRow?.tier);
-      const orgCredits = orgRow?.credits ?? null;
-
-      // Check concurrency FIRST — don't dequeue if no slot available
-      await checkRunConcurrencyLimit(orgId, orgTier, tx);
-
-      // Fetch all queue entries for this org (ordered FIFO).
-      // Typically 0-2 entries; iterating in-transaction to skip
-      // already-processed runs without releasing the advisory lock.
-      // JOIN agentRuns to read modelProvider for credit check.
-      const rows = await tx.execute<{
-        run_id: string;
-        encrypted_params: string | null;
-        model_provider: string | null;
-      }>(
-        sql`SELECT q.run_id, q.encrypted_params, r.model_provider
-         FROM agent_run_queue q
-         JOIN agent_runs r ON r.id = q.run_id
-         WHERE q.org_id = ${orgId}
-         ORDER BY q.created_at ASC`,
-      );
-
-      for (const row of rows.rows) {
-        // Delete queue entry
-        await tx
-          .delete(agentRunQueue)
-          .where(eq(agentRunQueue.runId, row.run_id));
-
-        // Check credits for VM0 provider runs
-        if (orgCredits !== null && orgCredits <= 0) {
-          const isVm0 = await resolveIsVm0Provider(
-            row.model_provider,
-            orgId,
-            tx,
-          );
-          if (isVm0) {
-            await transitionRunStatus(
-              row.run_id,
-              {
-                status: "failed",
-                error: insufficientCredits().message,
-                completedAt: new Date(),
-              },
-              ["queued"],
-              tx,
-            );
-            log.debug(`Run ${row.run_id} failed credit check, skipping`);
-            continue;
-          }
-        }
-
-        // Update run status — fails silently if run was already cancelled/failed
-        const [updated] = await tx
-          .update(agentRuns)
-          .set({ status: "pending", lastHeartbeatAt: new Date() })
-          .where(
-            and(eq(agentRuns.id, row.run_id), eq(agentRuns.status, "queued")),
-          )
-          .returning({ createdAt: agentRuns.createdAt });
-
-        if (!updated) {
-          // Run was cancelled/failed between enqueue and drain — skip it
-          log.debug(`Run ${row.run_id} already processed, skipping`);
-          continue;
-        }
-
-        log.debug(`Dequeued run ${row.run_id} for org ${orgId}`);
-        return {
-          runId: row.run_id,
-          createdAt: updated.createdAt,
-          encryptedParams: row.encrypted_params,
-        };
-      }
-
-      return undefined; // Queue empty (all entries were already processed)
-    });
-  } catch (error) {
-    if (isConcurrentRunLimit(error)) {
-      // No slot available — nothing was dequeued
-      return undefined;
-    }
-    throw error;
-  }
-}
-
-/**
- * Resolve whether the effective model provider is VM0.
- * Queries the model_providers table within the provided transaction
- * to maintain advisory lock guarantees.
- */
-async function resolveIsVm0Provider(
-  modelProvider: string | null,
-  orgId: string,
-  db: typeof globalThis.services.db,
-): Promise<boolean> {
-  if (modelProvider === "vm0") return true;
-  if (modelProvider && modelProvider !== "vm0") return false;
-  // null → check org default within the transaction
-  const [defaultProvider] = await db
-    .select({ type: modelProviders.type })
-    .from(modelProviders)
-    .where(
-      and(
-        eq(modelProviders.orgId, orgId),
-        eq(modelProviders.userId, ORG_SENTINEL_USER_ID),
-        eq(modelProviders.isDefault, true),
-      ),
-    )
-    .limit(1);
-  return defaultProvider?.type === "vm0";
-}
-
-/**
- * Clean up expired queue entries.
- * Marks associated runs as "timeout" and deletes the queue records.
- */
-export async function cleanupExpiredQueueEntries(): Promise<number> {
-  const now = new Date();
-
-  // Delete expired entries and collect their run IDs in one query
-  const deleted = await globalThis.services.db
-    .delete(agentRunQueue)
-    .where(lt(agentRunQueue.expiresAt, now))
-    .returning({ runId: agentRunQueue.runId });
-
-  if (deleted.length === 0) {
-    return 0;
-  }
-
-  const runIds = deleted.map((e) => e.runId);
-
-  // Mark associated runs as timeout (only if still queued)
-  await globalThis.services.db
-    .update(agentRuns)
-    .set({
-      status: "timeout",
-      completedAt: now,
-      error: "Queued run expired (exceeded queue TTL)",
-    })
-    .where(and(inArray(agentRuns.id, runIds), eq(agentRuns.status, "queued")));
-
-  log.debug(`Cleaned up ${deleted.length} expired queue entries`);
-  return deleted.length;
-}
-
-/**
- * Drain queues for orgs that have queued runs but no active runs.
- * Used as a cron fallback in case completion webhooks miss the drain.
- *
- * @param dispatch - Dispatcher function (injected to avoid circular dependency)
- */
-export async function drainStaleQueues(
-  dispatch: QueuedRunDispatcher,
-): Promise<number> {
-  const staleThreshold = new Date(Date.now() - PENDING_RUN_TTL_MS);
-
-  // Find distinct orgs with queued runs (orgId is denormalized on queue table)
-  const orgsWithQueued = await globalThis.services.db
-    .selectDistinct({ orgId: agentRunQueue.orgId })
-    .from(agentRunQueue);
-
-  let drained = 0;
-
-  for (const { orgId } of orgsWithQueued) {
-    // Check if org has any active runs (same logic as checkRunConcurrencyLimit)
-    const [result] = await globalThis.services.db
-      .select({ count: count() })
-      .from(agentRuns)
-      .where(
-        and(
-          eq(agentRuns.orgId, orgId),
-          or(
-            eq(agentRuns.status, "running"),
-            and(
-              eq(agentRuns.status, "pending"),
-              gt(agentRuns.createdAt, staleThreshold),
-            ),
-          ),
-        ),
-      );
-
-    const activeCount = Number(result?.count ?? 0);
-    if (activeCount === 0) {
-      log.debug(`Draining stale queue for org ${orgId}`);
-      await drainOrgQueue(orgId, dispatch);
-      drained++;
-    }
-  }
-
-  return drained;
-}
-
-const VALID_ORG_TIERS = new Set<string>(["free", "pro", "team"]);
-
-/** Parse a raw tier string into OrgTier, defaulting to "free". */
-function parseOrgTier(raw: string | undefined): OrgTier {
-  if (raw && VALID_ORG_TIERS.has(raw)) return raw as OrgTier;
-  return "free";
-}
-
-async function markQueuedRunFailed(
-  runId: string,
-  errorMessage: string,
-): Promise<void> {
-  await transitionRunStatus(
-    runId,
-    {
-      status: "failed",
-      error: errorMessage,
-      completedAt: new Date(),
-    },
-    ["queued", "pending"],
-  );
-}
