+++ turbo/apps/web/src/lib/zero/zero-preflight.ts
@@ -1,0 +1,237 @@
+import { eq, and, count, gt, or } from "drizzle-orm";
+import { env } from "../../env";
+import { agentRuns } from "../../db/schema/agent-run";
+import { orgMetadata } from "../../db/schema/org-metadata";
+import { orgMembersMetadata } from "../../db/schema/org-members-metadata";
+import { modelProviders } from "../../db/schema/model-provider";
+import {
+  concurrentRunLimit,
+  insufficientCredits,
+  noModelProvider,
+} from "../errors";
+import { ORG_SENTINEL_USER_ID } from "../org/org-sentinel";
+import { logger } from "../logger";
+import { MODEL_PROVIDER_ENV_VARS } from "../run/build-context";
+import type { Database } from "../../types/global";
+import type { AgentComposeYaml } from "../../types/agent-compose";
+import type { OrgTier } from "@vm0/core";
+
+const log = logger("zero:preflight");
+
+// Defense-in-depth: exclude pending runs older than this from concurrency check.
+// The cleanup-sandboxes cron job already transitions pending runs to "timeout" after 5 minutes,
+// so this TTL only matters if the cron job fails to run.
+export const PENDING_RUN_TTL_MS = 15 * 60 * 1000; // 15 minutes
+
+/** Concurrent run limits by org tier */
+const TIER_CONCURRENCY_LIMITS: Record<OrgTier, number> = {
+  free: 1,
+  pro: 2,
+  team: 5,
+};
+
+function getConcurrencyLimitForTier(tier: OrgTier): number {
+  return TIER_CONCURRENCY_LIMITS[tier];
+}
+
+/**
+ * Get the effective concurrency limit for an org tier.
+ * Tier-based limit is the baseline; env var acts as a global cap.
+ * Returns 0 for unlimited.
+ */
+export function getEffectiveConcurrencyLimit(orgTier: OrgTier): number {
+  const tierLimit = getConcurrencyLimitForTier(orgTier);
+  const envCap = env().CONCURRENT_RUN_LIMIT_CAP;
+  if (envCap === 0) return 0;
+  if (envCap !== undefined && !isNaN(envCap))
+    return Math.min(tierLimit, envCap);
+  return tierLimit;
+}
+
+/**
+ * Check if org has reached concurrent run limit
+ *
+ * @param orgId Clerk org ID to check
+ * @param orgTier Org tier for tier-based limit (default: "free")
+ * @param db Optional database instance (for use within transactions)
+ * @throws ConcurrentRunLimitError if limit exceeded
+ */
+export async function checkRunConcurrencyLimit(
+  orgId: string,
+  orgTier: OrgTier = "free",
+  db?: Database,
+): Promise<void> {
+  const effectiveLimit = getEffectiveConcurrencyLimit(orgTier);
+
+  // Skip check if limit is 0 (no limit)
+  if (effectiveLimit === 0) {
+    return;
+  }
+
+  const queryDb = db ?? globalThis.services.db;
+
+  // Count active runs: all "running" runs + "pending" runs within TTL
+  const staleThreshold = new Date(Date.now() - PENDING_RUN_TTL_MS);
+
+  const [result] = await queryDb
+    .select({ count: count() })
+    .from(agentRuns)
+    .where(
+      and(
+        eq(agentRuns.orgId, orgId),
+        or(
+          eq(agentRuns.status, "running"),
+          and(
+            eq(agentRuns.status, "pending"),
+            gt(agentRuns.createdAt, staleThreshold),
+          ),
+        ),
+      ),
+    );
+
+  const activeRunCount = Number(result?.count ?? 0);
+
+  if (activeRunCount >= effectiveLimit) {
+    log.debug(
+      `Org ${orgId} has ${activeRunCount} active runs, limit is ${effectiveLimit}`,
+    );
+    throw concurrentRunLimit();
+  }
+}
+
+/**
+ * Check if the org has sufficient credits for a VM0 provider run.
+ *
+ * Only blocks runs using the VM0 managed provider (where the platform
+ * bears API costs). Runs using the org's own API key are not affected.
+ *
+ * Uses lazy resolution: the org default provider is only queried
+ * when modelProvider is null AND credits are already depleted.
+ * The query runs within the caller's transaction to preserve
+ * advisory lock guarantees.
+ */
+export async function checkOrgCredits(
+  orgId: string,
+  userId: string,
+  modelProvider: string | null | undefined,
+  db?: Database,
+): Promise<void> {
+  const queryDb = db ?? globalThis.services.db;
+
+  // Explicit non-VM0 provider — skip check entirely
+  if (modelProvider && modelProvider !== "vm0") {
+    return;
+  }
+
+  // Determine if this is a VM0 run
+  let isVm0 = modelProvider === "vm0";
+
+  if (!isVm0 && !modelProvider) {
+    // Resolve org default provider to determine if this is a VM0 run
+    const [defaultProvider] = await queryDb
+      .select({ type: modelProviders.type })
+      .from(modelProviders)
+      .where(
+        and(
+          eq(modelProviders.orgId, orgId),
+          eq(modelProviders.userId, ORG_SENTINEL_USER_ID),
+          eq(modelProviders.isDefault, true),
+        ),
+      )
+      .limit(1);
+    isVm0 = defaultProvider?.type === "vm0";
+  }
+
+  // Per-member credit cap check — only for VM0 runs
+  if (isVm0) {
+    const [memberRow] = await queryDb
+      .select({ creditEnabled: orgMembersMetadata.creditEnabled })
+      .from(orgMembersMetadata)
+      .where(
+        and(
+          eq(orgMembersMetadata.orgId, orgId),
+          eq(orgMembersMetadata.userId, userId),
+        ),
+      )
+      .limit(1);
+
+    if (memberRow?.creditEnabled === false) {
+      throw insufficientCredits();
+    }
+  }
+
+  // Read credits from org_metadata
+  const [orgRow] = await queryDb
+    .select({ credits: orgMetadata.credits })
+    .from(orgMetadata)
+    .where(eq(orgMetadata.orgId, orgId))
+    .limit(1);
+
+  // No org row → treat as sufficient (new org, default 10000 credits)
+  if (!orgRow) {
+    return;
+  }
+
+  // Credits > 0 → sufficient for any provider
+  if (orgRow.credits > 0) {
+    return;
+  }
+
+  // Credits <= 0 and VM0 run — insufficient
+  if (isVm0) {
+    throw insufficientCredits();
+  }
+
+  // Effective provider is not VM0 — skip check
+}
+
+/**
+ * Pre-INSERT check: ensure the org has a model provider configured.
+ * Skips the check when:
+ * - The compose has explicit model provider env vars (e.g. ANTHROPIC_API_KEY)
+ * - An explicit modelProvider param is provided (validated later in build-context)
+ * - The framework doesn't use model providers (non-claude-code)
+ */
+export async function checkModelProviderConfigured(
+  orgId: string,
+  modelProvider: string | null | undefined,
+  composeContent: AgentComposeYaml,
+  db?: Database,
+): Promise<void> {
+  // Explicit modelProvider param provided — skip (will be validated in build-context)
+  if (modelProvider) return;
+
+  const queryDb = db ?? globalThis.services.db;
+
+  // Extract framework and environment from first agent
+  const firstAgent = composeContent.agents
+    ? Object.values(composeContent.agents)[0]
+    : undefined;
+  const framework = firstAgent?.framework || "claude-code";
+
+  // Only claude-code framework needs provider resolution
+  if (framework !== "claude-code") return;
+
+  // If compose has explicit model provider env vars, skip check
+  const hasExplicitConfig = MODEL_PROVIDER_ENV_VARS.some(
+    (v) => firstAgent?.environment?.[v] !== undefined,
+  );
+  if (hasExplicitConfig) return;
+
+  // Check if org has a default model provider
+  const [defaultProvider] = await queryDb
+    .select({ type: modelProviders.type })
+    .from(modelProviders)
+    .where(
+      and(
+        eq(modelProviders.orgId, orgId),
+        eq(modelProviders.userId, ORG_SENTINEL_USER_ID),
+        eq(modelProviders.isDefault, true),
+      ),
+    )
+    .limit(1);
+
+  if (!defaultProvider) {
+    throw noModelProvider();
+  }
+}
